Adjusted layer management for Paleogeography
